Server manifest & Multi-fork support (#5)
* Directory.Packages.props

* Delete database request logger.

Design is bad, I had to turn it off on our servers ages ago due to perf issues. I doubt anybody cares.

* Fix migrations running twice.

Apparently??

* Style settings in Rider

* Multi-fork support, test project, library

Content versions are now stored per fork.

Created a downloader project that serves as a test of the CDN. Added a Robust.Cdn.Lib project that sharable code will be moved to. In the future the launcher will be able to use this project instead.

A bunch of misc shit.

* Replace most custom Zstd streams with SharpZstd

* Disable request log in dev

* Make content ingestion work with new fork system, move to proper job scheduling library

* CDN fork publishing system

* Fix typo causing incorrect hash calculation

* Fix broken file get query

* Infrastructure to notify watchdogs of new published versions.

* Fix System.CommandLine reference being in Robust.Cdn.Lib

Should be in Robust.Cdn.Downloader

* Add user agent to HttpClients

* Private manifest fork support

* Automatically prune old manifest builds.

* Fancy build list page.

* Fix dockerfile